Fjernet synt kode og oppdatert hodejegerenConsumer
