+++ apps/testnorge-tp/build.gradle
@@ -65,7 +65,6 @@ dependencies {
 
     implementation 'io.springfox:springfox-swagger2:3.0.0'
     implementation 'io.springfox:springfox-swagger-ui:3.0.0'
-    implementation 'no.nav.registre:testnorge-consumers:1.12.0-RELEASE'
 
     runtimeOnly 'com.oracle.database.jdbc:ojdbc8:21.3.0.0'
