+++ apps/testnorge-tp/src/main/java/no/nav/registre/tp/consumer/HodejegerenConsumer.java
@@ -1,9 +1,8 @@
-package no.nav.registre.tp.consumer.rs;
+package no.nav.registre.tp.consumer;
 
 import io.micrometer.core.annotation.Timed;
 import lombok.extern.slf4j.Slf4j;
-import no.nav.registre.tp.consumer.rs.command.PostSaveHistorikkCommand;
-import no.nav.registre.tp.domain.TpSaveInHodejegerenRequest;
+import no.nav.registre.tp.consumer.command.GetLevendeIdenterCommand;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.client.ExchangeFilterFunction;
@@ -11,14 +10,15 @@ import org.springframework.web.reactive.function.client.WebClient;
 
 import java.util.List;
 
+
 @Component
 @Slf4j
-public class HodejegerenHistorikkConsumer {
+public class HodejegerenConsumer {
 
     private final WebClient webClient;
 
-    public HodejegerenHistorikkConsumer(@Value("${testnorge-hodejegeren.rest-api.url}") String hodejegerenServerUrl,
-                                        ExchangeFilterFunction metricsWebClientFilterFunction) {
+    public HodejegerenConsumer(@Value("${testnorge-hodejegeren.rest-api.url}") String hodejegerenServerUrl,
+                               ExchangeFilterFunction metricsWebClientFilterFunction) {
 
         this.webClient = WebClient.builder()
                 .baseUrl(hodejegerenServerUrl)
@@ -27,10 +27,10 @@ public class HodejegerenHistorikkConsumer {
     }
 
     @Timed(value = "tp.resource.latency", extraTags = {"operation", "hodejegeren"})
-    public List<String> saveHistory(
-            TpSaveInHodejegerenRequest request
+    public List<String> getLevende(
+            Long avspillergruppeId
     ) {
-        return new PostSaveHistorikkCommand(request, webClient).call();
+        return new GetLevendeIdenterCommand(avspillergruppeId, webClient).call();
     }
 
 }
