--- apps/testnorge-tp/src/main/java/no/nav/registre/tp/consumer/rs/TpSyntConsumer.java
@@ -1,50 +1,0 @@
-package no.nav.registre.tp.consumer.rs;
-
-import io.micrometer.core.annotation.Timed;
-import lombok.extern.slf4j.Slf4j;
-import no.nav.registre.tp.consumer.rs.command.GetSyntTpYtelserCommand;
-import no.nav.registre.tp.consumer.rs.credential.SyntTpProperties;
-import no.nav.registre.tp.database.models.TYtelse;
-import no.nav.testnav.libs.securitycore.domain.ServerProperties;
-import no.nav.testnav.libs.standalone.servletsecurity.exchange.TokenExchange;
-import org.springframework.stereotype.Component;
-import org.springframework.web.reactive.function.client.ExchangeFilterFunction;
-import org.springframework.web.reactive.function.client.ExchangeStrategies;
-import org.springframework.web.reactive.function.client.WebClient;
-
-import java.util.List;
-
-@Slf4j
-@Component
-public class TpSyntConsumer {
-
-    private final TokenExchange tokenExchange;
-    private final ServerProperties serviceProperties;
-    private final WebClient webClient;
-
-    public TpSyntConsumer(
-            SyntTpProperties syntTpProperties,
-            TokenExchange tokenExchange,
-            ExchangeFilterFunction metricsWebClientFilterFunction) {
-
-        this.serviceProperties = syntTpProperties;
-        this.tokenExchange = tokenExchange;
-        this.webClient = WebClient.builder()
-                .exchangeStrategies(ExchangeStrategies.builder()
-                        .codecs(configurer -> configurer
-                                .defaultCodecs()
-                                .maxInMemorySize(16 * 1024 * 1024))
-                        .build())
-                .baseUrl(syntTpProperties.getUrl())
-                .filter(metricsWebClientFilterFunction)
-                .build();
-    }
-
-    @Timed(value = "tp.resource.latency", extraTags = {"operation", "synt"})
-    public List<TYtelse> getSyntYtelser(
-            int numToGenerate
-    ) {
-        var token = tokenExchange.exchange(serviceProperties).block().getTokenValue();
-        return new GetSyntTpYtelserCommand(numToGenerate, token, webClient).call();
-    }
-}
