+++ apps/testnorge-tp/src/main/resources/application.properties
@@ -9,12 +9,7 @@ spring.datasource.driver-class-name=oracle.jdbc.driver.OracleDriver
 
 AAD_ISSUER_URI=https://login.microsoftonline.com/62366534-1ec3-4962-8869-9b5535279d0b
 
-testnorge-hodejegeren.rest-api.url=https://testnorge-hodejegeren.dev.intern.nav.no/api
-
-consumers.synt-tp-gcp.url=https://synthdata-tp-gcp.dev.intern.nav.no
-consumers.synt-tp-gcp.cluster=dev-gcp
-consumers.synt-tp-gcp.namespace=dolly
-consumers.synt-tp-gcp.name=synthdata-tp-gcp
+testnorge-hodejegeren.rest-api.url=https://testnorge-hodejegeren.dev.intern.nav.no
 
 management.endpoints.enabled-by-default=true
 management.endpoints.web.base-path=/internal
