+++ apps/testnorge-tp/src/test/resources/application-test.properties
@@ -1,4 +1,4 @@
-testnorge-hodejegeren.rest-api.url=http://localhost:${wiremock.server.port:0}/hodejegeren/api
+testnorge-hodejegeren.rest-api.url=http://localhost:${wiremock.server.port:0}/hodejegeren
 
 spring.datasource.url=jdbc:h2:mem:testdb-q2
 spring.datasource.driverClassName=org.h2.Driver
@@ -21,7 +21,3 @@ AAD_ISSUER_URI=http://localhost:${wiremock.server.port:0}/aad
 azure.app.client.id: dummy
 azure.app.client.secret: dummy
 
-consumers.synt-tp-gcp.url=http://localhost:${wiremock.server.port:0}/synt-tp
-consumers.synt-tp-gcp.cluster=dummy
-consumers.synt-tp-gcp.namespace=dummy
-consumers.synt-tp-gcp.name=synthdata-tp-gcp
